move lenses to own module; import microlens in tests only
