+++ package.yaml
@@ -69,6 +69,7 @@ tests:
       - doctest >= 0.9 && < 0.12
       - Glob >= 0.7 && < 0.8
       - yaml == 0.8.*
+      - microlens
 
 benchmarks:
   bench:
